build(webpack): Dynamically load Example Display
The example display component loads a lot of dependencies:
`js-beautify`, `vue-template-compiler`, `highlight.js` and `babel`.

By loading it asynchronously we are able to save a lot on the initial
load. Additionally we only load the syntax highlighting for `html`,
`css` and `javascript` in order to save a lot of unnecessary JS.